safari-600.7-branch/LayoutTests/http/tests/xmlhttprequest/timeout/xmlhttprequest-timeout-synconmain-expected.txt

CONSOLE MESSAGE: XMLHttpRequest.timeout cannot be set for synchronous HTTP(S) requests made from the window context.
CONSOLE MESSAGE: Synchronous XMLHttpRequests must not have a timeout value set.
Description

This test validates that the XHR2 timeout property behaves as expected in async cases in document (i.e. non-worker) context.


PASS XHR2 Timeout Property Tests 
PASS XHR2 Timeout Property Tests 1 
